dokumentacja cd

+++ src/main/java/com/example/survey/controller/UsersController.java
@@ -3,6 +3,7 @@ package com.example.survey.controller;
 import com.example.survey.model.Users;
 import com.example.survey.repository.UsersRepository;
 import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springdoc.core.annotations.ParameterObject;
 import org.springframework.http.HttpStatus;
@@ -21,9 +22,22 @@ public class UsersController {
 
     @DeleteMapping("/delete/{user_id}")
     @Operation(
-            description = "Delete a specific User"
+            description = "Delete a specific User",
+            responses = {
+                    @ApiResponse(
+                            description = "Succes",
+                            responseCode = "200"
+                    ),
+                    @ApiResponse(
+                            description = "Cant find user",
+                            responseCode = "204"
+                    )
+            }
     )
     public ResponseEntity<HttpStatus> deleteUserById(@PathVariable Integer user_id) {
+        if (usersRepository.findById(user_id) == null) {
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
         usersRepository.deleteById(Long.valueOf(user_id));
         return new ResponseEntity<>(HttpStatus.OK);
     }
@@ -40,7 +54,7 @@ public class UsersController {
 
     @GetMapping("/login")
     @Operation(
-            description = "Check if username and password are in the database"
+            description = "Check if user with given username and password are in the database and return his ID"
     )
     public ResponseEntity<Integer> getUser(@ParameterObject Users users) {
         Users newUsers = usersRepository.findByUsernameAndPassword(users.getUsername(), users.getPassword());
@@ -50,12 +64,11 @@ public class UsersController {
         return new ResponseEntity<>(newUsers.getId(), HttpStatus.OK);
     }
 
-    @PostMapping("/edit/{user_id}")
+    @PostMapping("/edit")
     @Operation(
             description = "Update user"
     )
-    public ResponseEntity<HttpStatus> updateUser(@RequestBody Users users, @PathVariable int user_id) {
-        users.setId(user_id);
+    public ResponseEntity<HttpStatus> updateUser(@RequestBody Users users) {
         usersRepository.save(users);
         return new ResponseEntity<>(HttpStatus.OK);
     }
